Revert "improve README, re-organize functions in run.py"
