+++ agent/__init__.py
@@ -1,8 +1,1 @@
-from .agent import (
-    Agent,
-    PromptAgent,
-    TeacherForcingAgent,
-    construct_agent,
-)
-
-__all__ = ["Agent", "TeacherForcingAgent", "PromptAgent", "construct_agent"]
+from .agent import *
